Start the non-interactive docker run attached, so that logs will just show. No explaining how to show the logs, as they'll just be visible to the user.
Make running it simpler - catching errors and giving the user guidance
when different `docker run` command line options are necessary, like
when needing to provide a qt.io build of Qt, or when the user didn't
provide options for showing the app on on the host.

Improve the docs in this regard, letting the inexperienced user know
they can just try the first `docker run` command I supply, then
following any guidance I provide at the end of that to show what to try
next.